requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pandas==2.2.0
requests==2.29.0
streamlit==1.31.0
numpy==1.26.3
openai==2.15.0
openai[embeddings]==2.15.0
google-generativeai==0.7.2
tiktoken
python-jose==3.3.0
customerio==2.1